Fix for `WilsonNF.calcCosWeightedVect`
In `WilsonNF.calcCosWeightedVect`, when the above-threshold neuron was at the start of the list (jj==0) or the end (jj==self.n-1), parabolic interpolation was not executed properly.

In the former case, the first arg to `parabolaInt` (`Rminus1`) should have been the last neuron in the list, i.e., `Neurons[0,self.n-1]`, but instead the above-threshold neuron `Neuron[0,jj]` was passed into the `parabolaInt` function (twice).

Similarly, in the latter case, the third arg to `parabolaInt` (`Rplus1`) should haave been the first neuron in the list, i.e., `Neurons[0,0]`, but instead the above-threshold neuron `Neurons[0,jj]` was passed into the `parabolaInt` function (twice), again.

In other words, the circular nature of the network was not being implemented properly in these edge cases. In this fix, modular arithmetic (e.g., `(jj+1)%self.n`) replaces the series of if-else statements to do the calculation in just one line.