Another fix for `WilsonNF.calcCosWeightedVect`
The final arg to `parabolaInt` (`Del`) was hard-coded as 15 degrees. When the number of nodes was 24, this was correct, but otherwise it was not.

A consequence of this error was that the perceived direction could have been significantly biased away from the node with the strongest activation when the number of nodes was increased from 24. For example, if the number of nodes was increased to 72, there should have been an angle difference of 360/72 = 5 degrees. However, the incorrect calculation always behaved like the angle difference between nodes was 15 degrees. Here's an example of the consequence: With the node activations (Rminus1, Rmax, Rplus1) = (0.1, 1, 0.95) and the angle of strongest activation MaxX = 0 degrees, the correct calculation of the perceived direction should be +2.2 degrees when the number of nodes is 72; however, with the coding error, the calculated perceived direction was +6.7 degrees instead, so the turn rate would have been over three times faster than it should be in this case.

When the number of nodes was decreased below 24, the perceived direction could have been biased towards the node with the strongest activation. Using the same example numbers as above and a reduced node count of 8, the correct calculation of the perceived direction should be +20.1 degrees. However, with the coding error, the calculated perceived direction was still only +6.7 degrees, so the turn rate would have been about three times slower than it should be in this case.