Add Agent.sensors() for iterating over sensors
The iteration scheme is based on code scattered and duplicated in several places, such as Simulation.updateSensors() and Simulation.updateSingleSensor() and the Extra Figures section. This new method is intended to yield authoritative information on each sensor's preferred angle relative to the head, its coordinates in the environment, and a sensor vector that is adjusted for the agent's orientation and which can be compared to a water vector. The method is not yet used.