Fix collisions with boundaries
There were two errors in AgentNF.checkInSquareBoundary. Instead of turning 180 degrees as intended, the agent's orientation vector was turned 90 degrees to the left, and its sensors were rotated 180 radians, or 10313.2 degrees (equivalent to approximately 127 degrees to the right). Not only did this lead to the agent not turning the correct amount, but its sensor placement and sensor detection vectors would become misaligned with its internal orientation vector.